+++ R/estimate_gps.R
@@ -5,35 +5,63 @@
 #' Estimates GPS value for each observation.
 #'
 #'
-#' @param y A vector of observed outcome variable (Y).
+#' @param Y A vector of observed outcome variable.
 #' @param w A vector of observed continuous exposure variable.
 #' @param c A data frame or matrix of observed covariates variable.
 #' @param pred.model The selected prediction model.
+#' @param internal.use If TRUE will return helper vectors as well. Otherwise,
+#'  will return original data + GPS value.
 #' @param ...  Additional arguments passed to the model.
 #'
 #' @return
-#' \code{matched_set}: The function returns a data.table saved the constructed
-#'  matched set by the proposed GPS matching approaches.
+#' The function returns a list of 6 objects according to the following order:
+#'   - Original data set + GPS values (Y, w, GPS, c)
+#'   - e_gps_pred
+#'   - e_gps_std_pred
+#'   - w_resid
+#'   - gps_mx (min and max of gps)
+#'   - w_mx (min and max of w).
+#' If \code{internal.use} is set to be FALSE, only originla data set + GPS will
+#' be returend.
 #'
 #' @export
-
-# Create matched set using GPS matching approaches
-EstimateGPS <- function(y,
+#'
+#' @examples
+#'
+#' m.d <- GenSynData(sample.size = 100)
+#' data.with.gps <- EstimateGPS(m.d$Y,
+#'                              m.d$treat,
+#'                              m.d[c("cf1","cf2","cf3","cf4","cf5","cf6")],
+#'                              pred.model = "sl",
+#'                              internal.use = FALSE,
+#'                              sl.lib = c("SL.xgboost","SL.earth","SL.gam",
+#'                                        "SL.ranger")
+#'                              )
+#'
+EstimateGPS <- function(Y,
                         w,
                         c,
                         pred.model,
+                        internal.use = TRUE,
                         ...){
 
-  e_gps <- TrainIt(Y = w, X = c, pred.model, ...)
-  e_gps_pred <- e_gps$SL.predict
-  e_gps_std <- TrainIt(Y = abs(w-e_gps_pred), X = c, pred.model, ...)
-  e_gps_std_pred <- e_gps_std$SL.predict
-  w_resid <- compute_resid(w,e_gps_pred,e_gps_std_pred)
-  gps <- compute_density(w_resid, w_resid)
-  w_mx <- compute_minmax(w)
-  gps_mx <- compute_minmax(gps)
-  dataset <- cbind(y,w,gps,c)
 
+  # Check passed arguments
+  CheckArgsEGPS(pred.model, ...)
 
-  return(list(dataset, e_gps_pred, e_gps_std_pred, w_resid, gps_mx, w_mx))
+  e.gps <- TrainIt(target = w, input = c, pred.model, ...)
+  e.gps.pred <- e.gps$SL.predict
+  e.gps.std <- TrainIt(target = abs(w-e.gps.pred), input = c, pred.model, ...)
+  e.gps.std.pred <- e.gps.std$SL.predict
+  w.resid <- ComputeResid(w,e.gps.pred,e.gps.std.pred)
+  gps <- ComputeDensity(w.resid, w.resid)
+  w.mx <- ComputeMinMax(w)
+  gps.mx <- ComputeMinMax(gps)
+  dataset <- cbind(Y,w,gps,c)
+
+  if (internal.use){
+    return(list(dataset, e.gps.pred, e.gps.std.pred, w.resid, gps.mx, w.mx))
+  } else {
+    return(dataset)
+  }
 }