+++ R/train_it.R
@@ -1,7 +1,7 @@
 #' Funtction to develop prediction model based on user's preferences.
 #'
-#' @param y A vector of target data. TODO: change these names to avoid confusion.
-#' @param x A vector, matrix, or dataframe of input data.
+#' @param target A vector of target data.
+#' @param input A vector, matrix, or dataframe of input data.
 #' @param model Prediction model algorithm.
 #'   - 'sl': SuperLearner The required parameters:
 #'     - *sl.lib*: a set of methods used for estimating target value (e.g.,
@@ -11,16 +11,13 @@
 #' @return
 #' prediction model
 #'
-#' @importFrom SuperLearner SuperLearner
 #' @keywords internal
 #'
-#'
-TrainIt <- function(Y, X, pred.model, ...) {
+TrainIt <- function(target, input, pred.model, ...) {
 
   # Passing packaging check() ----------------------------
   sl.lib <- NULL
   # ------------------------------------------------------
-
 
   dot_args <- list(...)
   arg_names <- names(dot_args)
@@ -30,11 +27,11 @@ TrainIt <- function(Y, X, pred.model, ...) {
   }
 
   if (pred.model == 'sl'){
-    pr_mdl <- SuperLearner(Y=Y, X=data.frame(X), SL.library=sl.lib)
+    pr_mdl <- SuperLearner::SuperLearner(Y=target, X=data.frame(input),
+                                         SL.library=sl.lib)
     return(pr_mdl)
   } else {
     stop(' This should not be raised. Something is wrong with CheckArgs
          function.')
   }
-
 }