+++ tests/testthat/test-compute_density.R
@@ -1,12 +1,11 @@
-test_that("multiplication works", {
+test_that("ComputeDensity works", {
 
   set.seed(223)
 
   a <- runif(100)
   b <- runif(100)
   a[10:12] <- NA
-  cd <- compute_density(a,b)
+  cd <- ComputeDensity(a,b)
   expect_equal(length(cd), 100)
   expect_equal(cd[10], 0.537411, tolerance = 0.0001)
-
 })