+++ tests/testthat/test-compute_outer.R
@@ -1,13 +1,13 @@
-test_that("multiplication works", {
+test_that("ComputeOuter function works fine.", {
 
   a1 <- 1:10
   a2 <- 1:20
-  r1 <- compute_outer(a1,a2,'-')
+  r1 <- ComputeOuter(a1,a2,'-')
 
   set.seed(892)
   b1 <- runif(200)
   b2 <- runif(500)
-  r2 <- compute_outer(b1,b2,'-')
+  r2 <- ComputeOuter(b1,b2,'-')
 
   expect_equal(dim(r1)[1], 10)
   expect_equal(dim(r1)[2], 20)