+++ tests/testthat/test-generate_synthetic_data.R
@@ -1,33 +1,29 @@
 test_that("Generated synthetic data length is correct.", {
-  s_data <- GenSynData()
-  expect_equal(nrow(s_data), 1000)
-  expect_equal(length(s_data), 8)
-  expect_true(is.data.frame(s_data))
+  s.data <- GenSynData()
+  expect_equal(nrow(s.data), 1000)
+  expect_equal(length(s.data), 8)
+  expect_true(is.data.frame(s.data))
 })
 
 test_that("Generated synthetic data type is correct.", {
-  s_data <- GenSynData()
-  expect_true(is.data.frame(s_data))
+  s.data <- GenSynData()
+  expect_true(is.data.frame(s.data))
 })
 
 test_that("Arguments provided for Generated synthetic data
            type is correct.", {
 
-  expect_error(GenSynData(sample_size = -10),
+  expect_error(GenSynData(sample.size = -10),
                regexp = "'sample_size' should be a positive ineteger numer." )
 })
 
 
 test_that(("Generated synthetic data is as expected."),{
 
-  data_1 <- GenSynData(sample_size = 200, seed=13221,
-                                    outcome_sd = 20)
-  data_2 <- GenSynData(sample_size = 1000, seed=734,
-                                    gps_spec = 2)
-  data_3 <- GenSynData(sample_size = 1250, seed=986,
-                                    outcome_sd = 8)
-  data_4 <- GenSynData(sample_size = 111, seed=1021,
-                                    gps_spec = 7)
+  data_1 <- GenSynData(sample.size = 200, seed=13221, outcome.sd = 20)
+  data_2 <- GenSynData(sample.size = 1000, seed=734, gps.spec = 2)
+  data_3 <- GenSynData(sample.size = 1250, seed=986, outcome.sd = 8)
+  data_4 <- GenSynData(sample.size = 111, seed=1021, gps.spec = 7)
 
   expect_equal(data_1$cf1[10] , -0.4354315, tolerance = 0.0001)
   expect_equal(data_1$cf5[79] , 0, tolerance = 0.0001)